refactored code and added main.ts
