+++ index.html
@@ -27,6 +27,7 @@
 		<footer>
 			<div id="footer"></div>
 		</footer>
+		<script type="module" src="/src/components/ui/header.ts"></script>
 		<script type="module" src="/src/app.ts"></script>
 	</body>
 </html>
