+++ src/components/tutorial/parts/tutorial-parts.ts
@@ -1,10 +1,22 @@
 import { introContent } from "../content/introContent";
 import { welcomeContent } from "../content/welcomeContent";
-import { TutorialPart } from "../../../managers/TutorialManager";
+
+export interface ContentItem {
+	content: string;
+}
+export interface TutorialPart {
+	id: string;
+	title: string;
+	content: ContentItem[];
+	contentClass: string;
+	buttonText: string;
+	shouldTutNavShow: boolean;
+}
 
 export const tutorialParts: TutorialPart[] = [
 	{
 		id: "welcome",
+		title: "Welcome",
 		content: welcomeContent,
 		contentClass: "text-justify",
 		buttonText: "Start Learning",
@@ -12,6 +24,7 @@ export const tutorialParts: TutorialPart[] = [
 	},
 	{
 		id: "intro",
+		title: "Introduction",
 		content: introContent,
 		contentClass: "text-left",
 		buttonText: "",
