+++ src/managers/ContentManager.ts
@@ -1,14 +1,6 @@
-export interface ContentItem {
-	content: string;
-}
+import { ContentItem } from "../components/tutorial/parts/tutorial-parts";
 
 export class ContentManager {
-	static goToPrevContent(): any {
-		throw new Error("Method not implemented.");
-	}
-	static goToNextContent(): any {
-		throw new Error("Method not implemented.");
-	}
 	private contentItems: ContentItem[];
 	private currentIndex: number = 0;
 	private contentElementId: string;
@@ -31,12 +23,8 @@ export class ContentManager {
 		if (contentElement && this.contentItems.length > 0) {
 			contentElement.innerHTML = this.contentItems[this.currentIndex].content;
 		}
-		this.updateContentVisibility(
-			"backward",
-			"forward",
-			"cta-button",
-			"tut-nav"
-		);
+		this.updateContentNavigationVisibility("backward", "forward", "cta-button");
+		this.updatePartNavigationVisibility("tut-nav");
 	}
 
 	public goToNextContent(): void {
@@ -78,16 +66,14 @@ export class ContentManager {
 		}
 	}
 
-	public updateContentVisibility(
+	public updateContentNavigationVisibility(
 		backwardElementId: string,
 		forwardElementId: string,
-		ctaButtonId: string,
-		tutNavId: string
+		ctaButtonId: string
 	): void {
 		const backwardElement = document.getElementById(backwardElementId);
 		const forwardElement = document.getElementById(forwardElementId);
 		const ctaButton = document.getElementById(ctaButtonId);
-		const tutNav = document.getElementById(tutNavId);
 
 		if (backwardElement) {
 			backwardElement.style.display = this.currentIndex > 0 ? "block" : "none";
@@ -102,7 +88,10 @@ export class ContentManager {
 			ctaButton.style.display =
 				this.currentIndex === this.contentItems.length - 1 ? "block" : "none";
 		}
+	}
 
+	public updatePartNavigationVisibility(tutNavId: string): void {
+		const tutNav = document.getElementById(tutNavId);
 		if (tutNav) {
 			const isAtLastContent =
 				this.currentIndex === this.contentItems.length - 1;
